REFACTOR conferenceInfo section
