+++ src/pages/conference/info/coOrganizer/coOrganizerInfo/style.scss
