+++ src/pages/conference/info/conferenceInfo/style.scss
