+++ src/store/ducks/conference/info/coOrganizer/actions.js
@@ -1,0 +1,9 @@
+import types from './types';
+
+const toggleCoOrganizerFormModal = () => ({
+  type: types.TOGGLE_CO_ORGANIZER_FORM_MODAL,
+});
+
+export default {
+  toggleCoOrganizerFormModal,
+};
