+++ src/store/ducks/conference/info/coOrganizer/index.js
@@ -1,0 +1,6 @@
+import reducer from './reducers';
+
+export { default as conferenceCoOranizerTypes } from './types';
+export { default as conferenceCoOranizerActions } from './actions';
+
+export default reducer;
