+++ src/store/ducks/conference/info/coOrganizer/types.js
@@ -1,0 +1,6 @@
+const TOGGLE_CO_ORGANIZER_FORM_MODAL =
+  'cem-admin/conference/info/TOGGLE_CO_ORGANIZER_FORM_MODAL';
+
+export default {
+  TOGGLE_CO_ORGANIZER_FORM_MODAL,
+};
