+++ src/store/ducks/index.js
@@ -1,3 +1,4 @@
 export { default as auth } from './auth';
 export { default as jsonplaceholder } from './jsonplaceholder';
 export { default as conference } from './conference';
+export { default as conferenceCoOranizer } from './conference/info/coOrganizer';
